refactor: 회원가입 funnel 컴포넌트 구현
