+++ src/app/sign-up/sign-up.css.ts
@@ -1,0 +1,9 @@
+import { style } from "@vanilla-extract/css";
+
+export const pageWrapper = style({
+  display: "flex",
+  flexDirection: "column",
+
+  width: "100%",
+  height: "100%",
+});
