+++ src/components/sign-up/Info/info.css.ts
@@ -1,4 +1,20 @@
 import { style } from "@vanilla-extract/css";
+import { vars } from "@/styles/theme.css";
+
+export const title = style({
+  ...vars.fontStyles.title2_sb_22,
+
+  marginTop: "5.2rem",
+  paddingLeft: "1rem",
+});
+
+export const description = style({
+  ...vars.fontStyles.body2_m_14,
+  color: vars.color.grey_6,
+
+  marginTop: "0.5rem",
+  paddingLeft: "1rem",
+});
 
 export const userProfileFormWrapper = style({
   flex: 1,
