+++ src/components/sign-up/StepIcon/StepIcon.tsx
