+++ src/components/sign-up/StepIcon/stepIcon.css.ts
