+++ src/components/sign-up/Terms/AllAgreeCheckboxField/AllAgreeCheckboxField.tsx
