+++ src/components/sign-up/Terms/TermCheckboxField/TermCheckboxField.tsx
