+++ src/components/sign-up/Terms/TermCheckboxField/termCheckboxField.css.ts
