+++ src/components/sign-up/Terms/terms.css.ts
@@ -1,24 +1,28 @@
 import { style } from "@vanilla-extract/css";
+import { vars } from "@/styles/theme.css";
 
-export const univInfoFormWrapper = style({
+export const title = style({
+  ...vars.fontStyles.title2_sb_22,
+
+  marginTop: "5.2rem",
+  paddingLeft: "1rem",
+});
+
+export const termsFormWrapper = style({
   flex: 1,
 
   display: "flex",
   flexDirection: "column",
   justifyContent: "space-between",
+  gap: "2rem",
 });
 
-export const univInfoFieldWrapper = style({
+export const termCheckboxListContainer = style({
   display: "flex",
   flexDirection: "column",
-  gap: "2rem",
+  gap: "1rem",
 
-  marginTop: "3.5rem",
-});
-
-export const swapButton = style({
-  display: "flex",
-  flexShrink: 0,
+  marginTop: "2.7rem",
 });
 
 export const fixedButtonWrapper = style({
